[GL] Move TINInterface to GeoLib/IO.
